+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/GeoShapeMapper.java
@@ -23,9 +23,12 @@ import com.stratio.cassandra.lucene.util.GeospatialUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.lucene.index.IndexableField;
 import org.apache.lucene.search.SortField;
+import org.apache.lucene.spatial.SpatialStrategy;
+import org.apache.lucene.spatial.composite.CompositeSpatialStrategy;
 import org.apache.lucene.spatial.prefix.RecursivePrefixTreeStrategy;
 import org.apache.lucene.spatial.prefix.tree.GeohashPrefixTree;
 import org.apache.lucene.spatial.prefix.tree.SpatialPrefixTree;
+import org.apache.lucene.spatial.serialized.SerializedDVStrategy;
 
 import java.util.Arrays;
 import java.util.Collections;
@@ -36,14 +39,14 @@ import static com.stratio.cassandra.lucene.util.GeospatialUtilsJTS.geometry;
 
 /**
  * A {@link Mapper} to map geographical shapes represented according to the <a href="http://en.wikipedia.org/wiki/Well-known_text">
- * Well Known Text (WKT)</a> format.
+ * Well Known Text (WKT)</a> format. <p> This class depends on <a href="http://www.vividsolutions.com/jts">Java Topology
+ * Suite (JTS)</a>. This library can't be distributed together with this project due to license compatibility problems,
+ * but you can add it by putting <a href="http://search.maven.org/remotecontent?filepath=com/vividsolutions/jts-core/1.14.0/jts-core-1.14.0.jar">jts-core-1.14.0.jar</a>
+ * into Cassandra lib directory. <p> Pole wrapping is not supported.
  *
- * This class depends on <a href="http://www.vividsolutions.com/jts">Java Topology Suite (JTS)</a>. This library can't
- * be distributed together with this project due to license compatibility problems, but you can add it by putting <a
- * href="http://search.maven.org/remotecontent?filepath=com/vividsolutions/jts-core/1.14.0/jts-core-1.14.0.jar">jts-core-1.14.0.jar</a>
- * into Cassandra lib directory.
- *
- * Pole wrapping is not supported.
+ * The indexing is based on a {@link CompositeSpatialStrategy} combining a geohash search tree in front of doc values.
+ * The search tree is used to quickly filtering according to a precision level, and the stored BinaryDocValues are used
+ * to achieve precision discarding false positives.
  *
  * @author Andres de la Pena {@literal <adelapena@stratio.com>}
  */
@@ -56,7 +59,7 @@ public class GeoShapeMapper extends SingleColumnMapper<String> {
     public final int maxLevels;
 
     /** The spatial strategy for radial distance searches. */
-    public final RecursivePrefixTreeStrategy strategy;
+    public final SpatialStrategy strategy;
 
     /** The sequence of transformations to be applied to the shape before indexing. */
     public final List<GeoTransformation> transformations;
@@ -67,8 +70,10 @@ public class GeoShapeMapper extends SingleColumnMapper<String> {
      * @param field the name of the field
      * @param column the name of the column
      * @param validated if the field must be validated
-     * @param maxLevels the maximum number of levels in the tree
-     * @param transformations the sequence of operations to be applied to the specified shape
+     * @param maxLevels the maximum number of precision levels in the search tree. False positives will be discarded
+     * using stored doc values, so this doesn't mean precision lost. Higher values will produce few false positives to
+     * be post-filtered, at the expense of creating many terms in the search index, specially with large polygons.
+     * @param transformations the sequence of operations to be applied to the indexed shapes
      */
     public GeoShapeMapper(String field,
                           String column,
@@ -85,26 +90,21 @@ public class GeoShapeMapper extends SingleColumnMapper<String> {
 
         this.maxLevels = GeospatialUtils.validateGeohashMaxLevels(maxLevels);
         SpatialPrefixTree grid = new GeohashPrefixTree(CONTEXT, this.maxLevels);
-        strategy = new RecursivePrefixTreeStrategy(grid, field);
 
-        this.transformations = (transformations == null) ? Collections.emptyList() : transformations;
+        RecursivePrefixTreeStrategy indexStrategy = new RecursivePrefixTreeStrategy(grid, field);
+        SerializedDVStrategy geometryStrategy = new SerializedDVStrategy(CONTEXT, field);
+        strategy = new CompositeSpatialStrategy(field, indexStrategy, geometryStrategy);
+
+        this.transformations = transformations == null ? Collections.emptyList() : transformations;
     }
 
     /** {@inheritDoc} */
     @Override
     public List<IndexableField> indexableFields(String name, String value) {
-
-        // Parse shape
         JtsGeometry shape = geometry(value);
-
-        // Apply transformations
-        if (transformations != null) {
-            for (GeoTransformation transformation : transformations) {
-                shape = transformation.apply(shape);
-            }
+        for (GeoTransformation transformation : transformations) {
+            shape = transformation.apply(shape);
         }
-
-        // Return strategy fields
         return Arrays.asList(strategy.createIndexableFields(shape));
     }